requirements.txt
click
numpy
cloud-volume
tinybrain
Pillow